+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * SQL: Change max primary key in DDL to 767 to fix MySQL
 * Elasticsearch: use an inline Groovy script to efficiently delete a property; also fixes query bugs
 * ACCUMULO: fix `markPropertyVisible` for edge properties
+* ACCUMULO: fix saving mutation with the save key/name/visibility properties with different timestamps
 
 # v2.2.10